refactor(gsd-db): extract db/engine.ts — connection + lifecycle + schema keystone
Phase A, commit 1 of the single-writer-layer refactor (CONTEXT.md decision).

Moves connection ownership (currentDb + workspace cache + open-state),
lifecycle (open/close/refresh/vacuum/checkpoint/snapshot), schema init +
migrations, and the BEGIN/COMMIT transaction primitives out of the 3,644-line
gsd-db.ts into db/engine.ts. The shared handle is now read through
getDb()/getDbOrNull(); writers/queries that closed over `currentDb` call the
accessor instead (behaviour-identical — existing guards keep their throw/return).

gsd-db.ts keeps the write/read wrappers and re-exports the engine surface
(`export *`), so every `from "./gsd-db.js"` import is unchanged. The
single-writer invariant test allowlists engine.ts (it legitimately owns DDL +
transaction control).

Verified: typecheck:extensions clean; single-writer-invariant + db-access
guardrails + db transaction/migration/cache/authority + compaction-snapshot
tests green.
